Admin PocketBase client for play updating
