+++ src/routes/(api)/api/modify/levelpack/+server.ts
@@ -1,5 +1,6 @@
 import type { RequestHandler } from "@sveltejs/kit";
-import { BAD, DENIED, isAdmin, OK, NOT_FOUND, MY_BAD } from "$lib/misc";
+import { isAdmin } from "$lib/misc";
+import { DENIED, MY_BAD, BAD, OK, NOT_FOUND } from "$lib/server/misc";
 import { ModifyLevelpackSchema } from "$lib/parse";
 import { getLevelpackByIdWithLevels } from "$lib/talk/get";
 import {
